Reimplement "set/clear inspection option" functionality using the new IntelliJ API #SCL-20513 Fixed
The new API provides a quickfix class for setting or clearing a boolean property in an inspection which allowed me to delete DontReportPublicDeclarationQuickFix
used in ScalaUnusedDeclarationInspection and easily create similar quickfixes for NonLocalReturnInspection and PrivateShadowInspection.

To work with our Scala plugin code, the new quickfix class needs to be wrapped in either `AbstractFixOnPsiElement` or `LocalQuickFixAndIntentionActionOnPsiElement`,
the latter in the case of `ScalaUnusedDeclarationInspection`. I think it's okay for now, the wrappers are thin, but we may think how to improve it in the future.