AbstractNotificationProvider: check if project is disposed before subscribing "ModuleRootChanged" event
This is mainly to workaround exceptions in logs during tests run.
It doesn't effect tests result but might be disturbing.
Long story shot, the issue is rooted to a long-existing issue in the platform: light projects (used for tests) are not properly disposed.
All core team is aware of it and for now there is no easy solution.
Also, unfortunately there is no any tracking issue for that.