move refactoring: fix NPE when moving class from file with non-type top level definitions in scala 3 #SCL-20395
The issue was that ScalaFileImpl.getClasses returned empty array even if there are classes in the file, but if if there are some non-type top-level definitions.
This was introduced in incorrect fix of #SCL-20395.
The method should always return classes if they exist in the file.
Some code in IntelliJ relies on that.

Remove ScalaFile.typeDefinitionsAndOthers, I don't think it has a place in API