Refactor CheapRefSearcher into a pipeline
Motivation: In the past 6 months we've added quite a lot of functionality
to what was once a relatively simple inspection for local unused declarations:
- global unused declaration inspection
- can-be-private inspection
- local implicit usage inspection
- etc.

The readability and maintainability of the associated code was starting to
suffer from these extra use cases. Especially the can-be-private inspection
increased the need for making an important abstraction explicit in the code,
which is the abstraction "CanExit", which models the question: "At which
point should I, the Cheap Reference Searcher, stop searching?"

Some other concepts that were introduced as explicit abstractions in the
code are "pipeline" and "search method".

All these concepts (pipeline, search method and can-exit) were present
before this refactoring, but they were buried and obfuscated by the way they
were implemented: lots of (nested) if-else.

In this commit we promote these concepts to first-class.

For more information on the introduced abstractions and their usage,
please read the docs in Search.scala.