refactoring tests: fix AbstractIntroduceVariableTestBase - it should use ScalaIntroduceVariableHandler.invoke directly
Previously it used some complicated logic, which was actually trying to emulate what is happening inside `ScalaIntroduceVariableHandler`. This was overcomplicated/redundant and didn't catch bugs in `ScalaIntroduceVariableHandler` itself.
In the new approach we propagate parameters for some internal refactoring behaviour via DataKey.
(This behaviour relies on user input in UI when refactoring is invoked in IDEA, for example selecting a name for a variable)