Enable BSP tracing only when selected
Previously, BSP tracing would only be enabled, but the tracer would not
write to disk when BSP tracing is de-selected in the IDE settings.

Enabling BSP tracing when the BSP messages are very large would
considerably slow down the import process, or completely prevent it,
because the BSP traces would still be produced, only to be later ignored
by the log writer.

This commit completely disables BSP tracing when the option is
de-selected in the IDE settings.