Simplify and improve SAM handling #SCL-20507
Added as part of SCL-20507 because I encountered an issue while
writing the below test:

test_trait_that_is_only_used_by_itself_via_method_parameter_type

The issue was that I like my tests to be surgical, so I avoid
allowAdditionalHighlights, and write the tests such that no
irrelevant highlights appear. Said test's trait happens to be
shaped such that its method is a SAM. The test was failing
because, despite the @unused annotation, the method was still
marked as unused. Continue reading to understand why.

The SAM use case is special because SAMs can be used without explicitly
referring to the method's name, so it makes sense to mark a SAM as used if
its containing class is used. However, what does not make sense, is to
mark a SAM as unused based exclusively on inspection of the containing
class, which is what the previous implementation did.

In this new implementation we mark a SAM as used whenever its containing
class is, or whenever the method is considered used, following the
same logic as for non-SAM abstract methods.