Allow ValueUsed to be based on hard pointer instead of SmartPsiElementPointer
Recently in our team more and more discussion started happening regarding some
flaky behaviour in highlighting. Every now and then a large amount of imports
and declarations suddenly are highlighted as unused, even though they are
used. Most of these declarations seem to be private, which leads us to suspect
ScalaRefCountHolder and friends (i.e. UsageTracker and ValueUsed) to be
involved. Since we don't have a reliable reproduction yet, we need to approach
the issue from multiple angles and do some elimination of potential causes.

This commit adds registry key scala.valueused.hardpointer, which, if set to
true, will result in hard pointers rather than SmartPsiElementPointers being
used for ValueUsed, as well as for ElementUsage.

The idea is to have some team members try out the hard pointer implementation
and see if this improves or alters the mentioned highlighting issue.