Reimplement targetCanBePrivate
While looking into a can-be-private false positive I noticed that
my previous targetCanBePrivate implementation was relatively hard
to understand, and it wasn't idiomatic.

Both should be addressed in this commit. In the previous
implementation an arbitrary tree-walk-up limit of up to 10 steps up
was imposed. I've decided to eliminate this limit, as we don't
seem to do something similar in many other places. We now simply
perform the required isAncestorOf checks, regardless of how nested
within a typedef a reference lies.