update sbt-idea-plugin to 3.19.4
- remove IntellijVersionUtils: it's moved to sbt-idea-plugin directly
- new version has improved fallback for intellijVersion in case some version is not publicly-available
E.g. if we have `intellijVersion = 231.1234.2` and it's a EAP-CANDIDATE version it will try downloading version 231.1234-EAP-CANDIDATE-SNAPSHOT.
And under the hood it contain IDEA with newer build: `231.1234.6` (though it still should be binary-compatible with 231.1234.2)
update sbt-idea-plugin to 3.19.3

- remove IntellijVersionUtils: it's moved to sbt-idea-plugin directly
- new version has improved fallback for intellijVersion in case some version is not publicly-available
E.g. if we have `intellijVersion = 231.1234.2` and it's a EAP-CANDIDATE version it will try downloading version 231.1234-EAP-CANDIDATE-SNAPSHOT.
And under the hood it contain IDEA with newer build: `231.1234.6` (though it still should be binary-compatible with 231.1234.2)
- remove custom CommonUltimate.upgradeIfNeeded: sbt-idea plugin can now handle it (see previous list item description)