sbt: don't show sbt-shell toolwindow when creating new sbt project, only ensure that it's available #SCL-20978 fixed
The issue is not reproduced reliably, though I am sure that the root cause lies in the modified code in ShowSbtShellAfterCreatingNewProject.
Originally this code was copied from `ExternalToolWindowManager` which ensures that ES toolwindow is shown. In case of sbt projects it is "sbt" tool window.
However "sbt-shell" is strictly speaking not a ES tool window because it doesn't implement `com.intellij.openapi.externalSystem.service.task.ui.AbstractExternalSystemToolWindowFactory`
So we don't actually need to show the tool window.
Showing it will lead to it's initialization which is not desired (see SCL-20978)
We only need to ensure that "sbt-shell" tool window available