drop legacy support for "inline" from scala-meta "macro paradise" plugin #SCL-21031 fixed
From chat with Mikhail Mutcianko:
Dmitrii Naumenko
Hey! Do you remember what was this commit about?
https://github.com/JetBrains/intellij-scala/commit/7585dd50b5203bf6f867f5d7594071b406ed1160
`f.getModifierList.findFirstChildByType(ScalaTokenTypes.kINLINE) != null`
At those times there was no Scala 3, and `inline` wasn't a modifier.
There was only annotation `@inline` and that's it.
This code leads to SCL-21031

Mikhail Mutcianko
It's a workaround because `inline` was a keyword added by macro paradise compiler plugin.
It wasn't official part of language and was parsed depending on whether compiler plugin was enabled or not ¯\_(ツ)_/¯

AFAIR this plugin was deprecated as well as scala meta in the compiler
I think you can simply remove this and everything which is about  *Meta.*