Convert HighlightingPass inspections to LocalInspectionTools #SCL-20531
Formerly, 3 inspections (var-could-be-val, unused-declaration and
can-be-private), were implemented via a rather arcane setup. I don't know
all the details and reasoning behind this setup, but I do know it was
based on ScalaRefCountHolder, which is old and flaky at best.

I don't see a reason to hold on to this setup. In fact, by implementing
these inspections as highlightingpasses we miss out on all the UX
optimizations provided by platform in LocalInspectionTools, which are:
highlight visible area first, display highlights incrementally rather
than in bulk, process suspicious elements (elements that were problematic
during a previous run) first, and let platform rearrange the order in
which inspections are run.

Some more reasons to sunset our previous approach for these 3 inspections:
* Follow platform recommendations and conventions.
* Code understandability.
* #SCL-20814 is now automagically fixed.