Implement Scala CLI parsing and syntax highlighting #SCL-21139
In this commit we introduce the first bits and pieces of Scala CLI
support, in the form of lexing, parsing and syntax highlighting Scala
CLI directives.

The approach is largely identical to our ScalaDoc support. Our
overarching lexer, _ScalaSplittingLexer, is responsible for separating
Scala code from ScalaDoc comments. This lexer knows exactly enough to
perform this separation, after which further lexeing is done by
_ScalaCoreLexer and _ScalaDocLexer.

In this commit we give _ScalaSplittingLexer just enough knowledge of
Scala CLI directives to capture them. Each directive is inserted into
the AST and PSI tree as a LazyParseablePsiElement. More fine-grained
lexing of these directives is performed by _ScalaCliLexer, after which
individual constituents of CLI directives start to appear as individual
tokens: the CLI prefix, the command -- of which at the moment there
exists only one type, "using" -- and an optional array of values.

After lexing, the actual parsing takes place. ScalaCliParser scans
lexemes produced by _ScalaCliLexer in order of occurrence, and
determines whether a given occurrence is congruent with the grammar of
Scala CLI directives. If yes, the parser simply continues to the next
token. Only if an occurrence does not follow the CLI grammar, the parser
reports this to the platform by calling `builder.error(...)`. Note that
in ScalaCliParser there is currently no recovery after such an error
occurred, and no further parsing will take place within the context of
that specific directive.

Once the Parser is done processing all available lexer tokens,
ScalaCliParserDefinition maps these tokens to PSI elements. Contrary to
our ScalaDoc implementation, in this commit we don't introduce PSI
element implementations for constituents of Scala CLI directives. The
requirements for these elements are not clear yet, nor are they
necessary to satisfy the requirements of SCL-21139. So the
`createElement` implementation in this ParserDefinition is trivial,
wrapping each token in a ASTWrapperPsiElement.

Moreover we introduce ScalaCliSyntaxHighlighter, which mostly speaks for
itself. Special attention must be paid, however, to the
`registerSelfStoppingLayer` call this commit adds to
ScalaSyntaxHighlighter. Without it, only the first occurrence of a CLI
directive within a file is highlighted correctly.