Fix getContainingClass of ScFunctionDefinitionImpl
The getContainingClass method is only used in JVM-facing contexts.
When such a context asks a Scala function what class contains it,
it should only reply with a synthetic module if this function is
in an implicit AnyVal class, and the function itself has
visibility that makes it an extension method -- public.

This commit specifically implements the visibility constraint.