[scalatest] show green gutter icons for individual tests in scala 3 code #SCL-21497 fixed
This change includes a workaround which uses a hardcoded mapping: "scala test base class fqn" -> finder instance.
It's only designed for idea232.release/x branches.
In branch idea233.x there will be a proper fix in the decompiler (Tasty reader).
After that fix finder annotations will be properly detected so this workaround is not needed.

#noport