Properly fix nested-if unnecessary parentheses inspection #SCL-21264 fixed
The earlier fix in f2257472 did not address the issue completely, i.e.
recursively.
It also introduced a regression by ignoring the inner if completely. The
regressed case was "inner if-tree ends with else", in which case one can
always remove parentheses without changing semantics. A concrete example
of this regressed case:

if (true) (if (true) println(1) else println(2)) else println(3)