Completely reimplemented ScalaControlFlowBuilder
I basically couldn't understand the previous implementation.
The pending mechanic was overly complicated and
because of all the mutability, it was also very difficult to
reason about.

The new implementation should be much more easy to understand.
Maybe :D

Because of the new, BETTER implementation many of the
test cases have changed slightly. Hope that is no problem

 #SCL-21679