Revert "[project view/structure view] use special icon for runnable objects" #SCL-13613 open
This reverts commit 5f2e1104cf0aa54807ee8128473e415118497d8a.

I reverted the commit because there are some questions regarding the performance.
This icon can be shown in the project view.
To check if the class inside is runnable in general case we need to check the entire hierarchy of the class.
This might be too heavy operation to run on all visible nodes in project view.
We first need to investigate the performance of this change