This was a test to not do the same calculations multiple time at the same time, but wait on the first to finish.
Our caching for certain computed properties works by checking an atomic var if the value has already been computed and taking that or doing the computation and save the result into the atomic var. If multiple threads determine hat a property has not been computed, they all start the computation, though only the first one gets to store its result.

This commit tries to make it so that only the first thread starts the computation and the rest wait for the first thread to finish. This should in theory save cpu load and may improve performance.

Unfortunately this behaviour might lead to deadlocking. That's why there is a deadlocking preventing mutex implementation.

From what I could see from the build server, this didn't bring any noticeable performance increase... so I'll leave it here for later investigation.