fixed MacOS one pixel shift of aligned inlay hints #SCL-21630 fixed
the problem was that, macos has fractional pixel coordinates,
which pose a problem when using swings graphics api.
We were calculating the margin of the inlay hints in integers,
but that lead to rounding errors.
Now we just set the start of the inlay hints to the same x value for
all hints in an alignment-group instead of finding a margin
that correctly calculates this within the renderer