ScTypeDefinitionImpl.getQualifiedName shouldn't return null for package object #KTIJ-24653
We violated contract of `getQualifiedName` contract.
It states that "null should be returned for anonymous and local classes and for type parameters"
We returned null for package object, but from the JVM perspective pakcage object is represented with a package with `package` in it's FQN.
The issue was found when discussing KTIJ-24653 Ilya Kirillov.

Plus, refactor ScTypeDefinitionImpl.getPackageName:
- avoid implicit parameters where they are not needed
- introduce "inner" recursive function, which doesn't pass "separator" all the time

Also related to #SCL-21894