[run configurations] fix error shown in application run configuration for objects extending App #SCL-22005 fixed
Technical details:
When Java operates with Scala Objects in some places, it actually works with an instance of PsiClassWrapper (org.jetbrains.plugins.scala.lang.psi.light.PsiClassWrapper)
When IntelliJ checks if the method has a main method to show an error in the Application Run Configuration, it uses `com.intellij.psi.util.PsiMethodUtil.findMainMethod(com.intellij.psi.PsiClass)` under the hood.
Previously it worked for Scala objects due to the default Java implementation of `com.intellij.psi.util.PsiMethodUtil.findMainMethod(com.intellij.psi.PsiMethod[], com.intellij.psi.PsiClass)`, but it didn't use our implementation of `JavaMainMethodProvider` (`ScalaMainMethodProvider.isApplicable` returned false for `PsiClassWrapper`)
After IDEA-339923 was fixed, it stopped working for Scala object because an extra check was added there:
`!inheritedStaticMainEnabled(containingClass)` which only returns true since Java 22 Preview. It's only executed when the containing class of the main method candidate was not equal to the original class:
`if (containingClass != null && containingClass != aClass) {`
But that was the actual case. Even though we returned a `ScFunctionWrapper`, its containing class wasn't equal to the original `PsiClassWrapper` instance, it returned a new instance.

Instead of relying on some java logic, which relies on some method wrappers, it makes sense to move the implementation `ScalaMainMethodProvider` and have more control on that.