Support export statements in extensions
#SCL-22266

With export statements being allowed inside extensions,
we now have to be extra careful when manipulating extension methods
in general. The problem is: we can no longer statically
determine if given ScFunction is an extension method or not,
as it depends on how it was resolved, what that means is
`fn.extensionMethodOwner == None` no longer guarantees that `fn`
cannot be used as an extension.