use recursion guard when building class signatures in ScalaPsiManager.SignatureCaches #SCL-22786
- Introduce registry controlling the behavior (ON by default):
`use.recursion.guard.for.signatures.cache`
- Use similar handling of classes as for ScFunction in CacheWithRecursionGuard/CachesUtil.handleRecursiveCall

I was investigating performance issues in the project from #SCL-22786.
I noticed that when analyzing files from the project,
it calculates signatures for the same classes many times, and it happens in recursion for the same arguments.
The project uses a lot of wildcard imports and has a lot of chaotic interdependencies between different classes.
I couldn't come up with a minimized example, though.