[revive old commits] [annotator] scala 3: show error for methods with `()` parameter clauses accessed without the empty argument list `()` #SCL-22178
Error message: "Method {0} must be called with () argument"

Before Scala 3 it was a warning, shown in `ParameterlessAccessInspection.EmptyParenMethod`.
Right now we just reuse the same logic from the inspection.
For that I also had to make `AddCallParenthesesQuickFix` to extend `LocalQuickFixAndIntentionActionOnPsiElement` in order it can be registered both in inspection and annotator

Also related: #SCL-22179, #SCL-18877