[cbh] don't register CompilerHighlightingEditorFocusListener on synthetic files #SCL-23885
Note, this doesn't fix any big memory leaks.
It seems that `org.jetbrains.plugins.scala.compiler.highlighting.CompilerHighlightingEditorFocusListenerRegisterer$MyDisposable` was just unlucky to hold references to too many editors during the "low memory" event when the report is generated. But it does it because other parts of the IDE didn't dispose of those editors.