[internal build] move dependency on scalac2Patches compiler plugin closer to the place where the plugin is used
Incremental compiler can drop caches for scalac2Patches, e.g. when switching the branch. But the incremental build can still start from the scalaIml module, not from the whole project rule. In this case the compilation will fail because the compiler plugin target/classes will be empty, and it's not recompiled because it wasn't the dependency of the scalaImpl