[editor/enter handler] don't commit the document after pressing Enter after `{` in AddUnitFunctionSignatureEnterHandler (#SCL-24362, #SCL-12690))
Comments to the change in `checkBlock2` method:
The OLD implementation checked for 2 new lines because it depended on 2 things
 1. The line was added in ScalaEnterAfterUnmatchedBraceHandler before AddUnitFunctionSignatureEnterHandler
 2. AddUnitFunctionSignatureEnterHandler invoked "commitDocument" before the main logic to make sure that the potential new line was visible in the new PSI structure (commit flushes the document changes to the PSI tree).

If we don't commit the document, we can check for just one new line symbol.

It's not directly related to the MAJOR freeze in SCL-24362.
However, I noticed it in the process, when measuring the action and decided to make the improvement in the same iteration.