[fix freezes] TaskRunnerWithLoadingProgress: remove redundant progress cancellation check #SCL-24416
See the line comments on this change for details.
In short, the extra cancellation checks on the UI thread are not needed because if the loading progress action was canceled (via Escape key or focus change), then the ReadAction is also canceled (thanks to `expireWith`) and the logic on the UI thread will never be invoked.