[run configurations, sbt] notify execution listeners when running sbt run configuration #SCL-24434 fixed
This is needed in order the "Before launch" feature works if the sbt run configuration is used as a dependent configuration. Otherwise, the main run configuration will never start after sbt task is finished.

Note, we don't need this when sbt run configuration is executed in "Debug" mode (org.jetbrains.sbt.runner.SbtDebugProgramRunner).
This is because in the Debug mode, an extra process is launched to attach to the sbt shell process.
And that process content descriptor is substituted in SbtDebugProgramRunner.createContentDescriptor and then sends started/terminated events.

(cherry picked from commit 094eab98a3a061135809ff07594a3de9243862ba)