fix SOE in implicit search with java cyclic raw types #SCL-24369 fixed
This is a continuation of SCL-19526
(related: https://github.com/JetBrains/intellij-scala/pull/620)
Please see the original ticket first for the details.

The original fix doesn't work when a raw type has type parameter with a bound that cyclically depends on itself.
For example, notice that `I<C>` uses raw type `C`, without type argument.
```java
abstract class C<T extends C<T>> implements I<C> { abstract void foo(); }
interface I<T> {}
```
During super types processing, the corresponding scala representation of `I<C>`
will be  `I[(C[T]) forSome {type T <: C[T]}]`
and the `C` is represented as `ScExistentialArgument.Deferred` that can't have a proper equality check.
The original fix of SCL-24369 fixed it for java raw types to an `ScAbstractType` instance.
However, the original fix preserved the bound of the type. But in this case the bound of T cyclically depends on `C[T]`.
And under the hood of the type object fields references it again references to a `ScExistentialArgument.Deferred`.
So the equality check for this new ScAbstractType instance will not work as well.
It means that when the `ScAbstractType` is processed further in the recursive cycle of the `collectParts` it can't detect the cycle leading to SOE.
In this fix, instead of preserving the original bounds we pretend them to be `Any` and `Nothing`
as if the code used `?`: `abstract class C<T extends C<T>> implements I<C<?>>`

(cherry picked from commit bcc90aa6bbdf6037a74aedb8b97d49bcaa843069)