[run configurations, sbt] [debugger fix] add a missing SbtDebugProgramRunner.canRun method that causes debugger to completely not work #SCL-24434 fixed, #SCL-24469 fixed
I forgot this method in the previous commit.
I leave the fix in a separate commit for transparency, in order it's clear what is different from the previous attempt to fix SCL-24434

Without this method it uses the default `GenericDebuggerRunner.canRun` implementation leading to more then 1 runner to be able to run any module-based configuration, breaking the debugger for ANY module-based run configuration.