[gutters, inheritors search, freeze] delegate class inheritors search to GotoImplementationHandler().navigateToImplementations #SCL-24511 fixed
The method runs the progress in the background and fills the list asynchronously while the new elements are found (you can test it by simply adding a Thread.sleep(300) in the `GotoImplementationHandler.ImplementationsUpdaterTask.run` method in a non-suspending breakpoint.

The previous implementation that we had didn't have much custom logic.
It used the `ClassInheritorsSearch.search` from the platform and modified the titles which seem redundant.
Now we use a single title with "Implementations" for traits and classes. Don't use "Subclasses" for non-traits. This is how it's done for Java for interfaces and classes.


(cherry picked from commit a43df7009bf836dc64972e3fef26007061580825)