[tests] cleanup ScalaCompilerTestdataHighlightingTest-based tests
doTest() method behavior was too implicit.
In the failing tests it was running single file/dir tests, and in non-failing it was running all tests in the directory. This was due to the differences of `filesToHighlight` method implementation. Now the data flow is more explicit + I introduced a dedicated method `doTestForFileOrDirectory` that makes the test cases more readable and self-explanatory