[sbt/libraries] save library maven coordinates (group/artifact/version) #SCL-24506
It's indirectly necessary for #SCL-24506 to detect version of uTest library, to conditionally choose some logic based on it (to detect or not to detect "class" test suites).
However, this change can affect some other functionality coming from IntelliJ platform automatically. It's hard to predict the full scope of it.

Related entities, involved under the hoo, in the platform
- com.intellij.java.library.MavenCoordinates
- com.intellij.java.library.LibraryWithMavenCoordinatesProperties
- com.intellij.externalSystem.ImportedLibraryProperties
- org.jetbrains.jps.model.library.JpsMavenRepositoryLibraryDescriptor (for JPS projects)

Also, I considered adding similar logic for BSP, but it's hard to implement right now