[types] fix type presentation for type alias definitions inside refinements #SCL-24691 fixed
Patch `compoundTypeText` in `ScalaTypePresentation.scala`.
When rendering members from the refinement, we need to pass the new type presentation context. This is important when we calculate `isElementResolvedFromContextWithoutQualifier` when we decide if we can drop the qualified name.

Before this fix, all the types inside the `compoundTypeText` used the original type presentation context passed to `typeText` method. The context was equal to the compound type definition place.
Take this example:
```scala
object Parsers {
  trait Builder

  val value1 = (??? : {
    type Builder = Parsers.Builder
  })
}
```
When `type Builder = Parsers.Builder` was rendered, it used the context of `value1`. As a result, the value of `isElementResolvedFromContextWithoutQualifier` was true because the `Builder` is present in `Parsers` object and it was resolved without the qualifier. But if we use the updated presentation context equal to the alias inside the compound type `type Builder = Parsers.Builder` we would see that the `Builder` is resolved to the type alias itself, not the `Parsers.Builder`. Thus we don't drop the qualifier.


(cherry picked from commit 021892a221e087ae9baecf53e81190f6c70e23f7)